Move tools downloading code to toolsWrangler
So it can be reused from import.
