+++ commands/upgradeagents.go
@@ -4,17 +4,12 @@
 package commands
 
 import (
-	"archive/tar"
 	"bytes"
-	"compress/gzip"
 	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path"
-	"strings"
 	"text/template"
 
 	"github.com/juju/cmd"
@@ -26,7 +21,6 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/juju/1.25-upgrade/juju2/api"
-	coretools "github.com/juju/1.25-upgrade/juju2/tools"
 )
 
 //go:generate go run ../juju2/generate/filetoconst/filetoconst.go agentUpgradeScript agent-upgrade.py agentupgrade_script.go 2017 commands
@@ -163,11 +157,10 @@ func (c *upgradeAgentsImplCommand) Run(ctx *cmd.Context) error {
 	}
 
 	// Get the tools from the controller.
-	client := utils.GetNonValidatingHTTPClient()
-	toolsURLPrefix := fmt.Sprintf("https://%s/tools/%s-", conn.Addr(), ver)
+	tw := newToolsWrangler(conn)
 	for _, seriesArch := range toolsNeeded.SortedValues() {
-		if err := c.getTools(ctx, client, ver, toolsURLPrefix, seriesArch); err != nil {
-			return errors.Annotatef(err, "downloading tools %s-%s", ver, seriesArch)
+		if err := tw.getTools(seriesArch); err != nil {
+			return errors.Trace(err)
 		}
 	}
 
@@ -182,35 +175,6 @@ func (c *upgradeAgentsImplCommand) Run(ctx *cmd.Context) error {
 		return errors.Trace(err)
 	}
 	return errors.Trace(reportResults(ctx, "upgrade", machines, results))
-}
-
-func (c *upgradeAgentsImplCommand) getTools(ctx *cmd.Context, client *http.Client, ver version.Number, toolsURLPrefix, seriesArch string) error {
-	toolsUrl := toolsURLPrefix + seriesArch
-	toolsVersion := version.MustParseBinary(ver.String() + "-" + seriesArch)
-
-	// Look to see if the directory is already there, if it is, assume
-	// that it is good.
-	downloadedToolsDir := path.Join(toolsDir, toolsVersion.String())
-	if _, err := os.Stat(downloadedToolsDir); err == nil {
-		fmt.Fprintf(ctx.Stdout, "%s exists\n", downloadedToolsDir)
-		return nil
-	}
-
-	fmt.Fprintf(ctx.Stdout, "Downloading tools: %s\n", toolsUrl)
-	resp, err := client.Get(toolsUrl)
-	if err != nil {
-		return errors.Annotate(err, "downloading tools")
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return errors.Errorf("bad HTTP response: %v", resp.Status)
-	}
-
-	err = UnpackTools(toolsDir, toolsVersion, resp.Body)
-	if err != nil {
-		return errors.Errorf("cannot unpack tools: %v", err)
-	}
-	return nil
 }
 
 func (c *upgradeAgentsImplCommand) saveMachines(machines []FlatMachine) error {
@@ -279,98 +243,12 @@ func (c *upgradeAgentsImplCommand) writeUpgradeScript(config *scriptConfig) (str
 	}
 	return scriptPath, nil
 }
-
-// UnpackTools reads a set of juju tools in gzipped tar-archive
-// format and unpacks them into the appropriate tools directory
-// within dataDir. If a valid tools directory already exists,
-// UnpackTools returns without error.
-func UnpackTools(dataDir string, toolsVersion version.Binary, r io.Reader) (err error) {
-	// Unpack the gzip file and compute the checksum.
-	zr, err := gzip.NewReader(r)
-	if err != nil {
-		return err
-	}
-	defer zr.Close()
-	f, err := ioutil.TempFile(os.TempDir(), "tools-tar")
-	if err != nil {
-		return err
-	}
-	_, err = io.Copy(f, zr)
-	if err != nil {
-		return err
-	}
-	defer os.Remove(f.Name())
-
-	// Make a temporary directory in the tools directory,
-	// first ensuring that the tools directory exists.
-	dir, err := ioutil.TempDir(toolsDir, "unpacking-")
-	if err != nil {
-		return err
-	}
-	defer removeAll(dir)
-
-	// Checksum matches, now reset the file and untar it.
-	_, err = f.Seek(0, 0)
-	if err != nil {
-		return err
-	}
-	tr := tar.NewReader(f)
-	for {
-		hdr, err := tr.Next()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return err
-		}
-		if strings.ContainsAny(hdr.Name, "/\\") {
-			return fmt.Errorf("bad name %q in tools archive", hdr.Name)
-		}
-		if hdr.Typeflag != tar.TypeReg {
-			return fmt.Errorf("bad file type %c in file %q in tools archive", hdr.Typeflag, hdr.Name)
-		}
-		name := path.Join(dir, hdr.Name)
-		if err := writeFile(name, os.FileMode(hdr.Mode&0777), tr); err != nil {
-			return errors.Annotatef(err, "tar extract %q failed", name)
-		}
-	}
-	// Write some metadata about the tools.
-	tools := &coretools.Tools{Version: toolsVersion}
-	toolsMetadataData, err := json.Marshal(tools)
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(path.Join(dir, toolsFile), []byte(toolsMetadataData), 0644)
-	if err != nil {
-		return err
-	}
-
-	// The tempdir is created with 0700, so we need to make it more
-	// accessable for juju-run.
-	err = os.Chmod(dir, 0755)
-	if err != nil {
-		return err
-	}
-
-	return os.Rename(dir, path.Join(toolsDir, toolsVersion.String()))
-}
-
 func removeAll(dir string) {
 	err := os.RemoveAll(dir)
 	if err == nil || os.IsNotExist(err) {
 		return
 	}
 	logger.Errorf("cannot remove %q: %v", dir, err)
-}
-
-func writeFile(name string, mode os.FileMode, r io.Reader) error {
-	f, err := os.OpenFile(name, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, mode)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	_, err = io.Copy(f, r)
-	return err
 }
 
 func seriesArch(machine FlatMachine) string {
